Update README.md
This commit is the final version of the inventions intended to be included in the Employment Agreement with Memphis Meats. Changes beyond this commit cannot be considered to be part of the original Employment Agreement. A URL to this list is included in Exhibit C of the employment Agreement. This is generated because of the lack of more than one line for writing in the Docusign system for this particular exhibit.